Move global to uop rewrite rule to the new rewrite folder
