+++ src/bloqade/qasm2/rewrite/__init__.py
@@ -1,2 +1,3 @@
+from .global_to_uop import GlobalToUOpRule as GlobalToUOpRule
 from .parallel_to_uop import ParallelToUOpRule as ParallelToUOpRule
 from .uop_to_parallel import UOpToParallelRule as UOpToParallelRule
